server: Rotas autenticadas

+++ server/src/routes/memories.ts
@@ -4,9 +4,18 @@ import { z } from 'zod'
 import { prisma } from "../lib/prisma"
 
 export async function memoriesRoutes(app: FastifyInstance) {
+  // Verificação de autenticação do usuário antes de cada execução de persistencia para este arquivo
+  app.addHook('preHandler', async (request) => {
+    // cancela a execução caso não tenha token válido by @fastify/jwt
+    await request.jwtVerify()
+  })
 
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -22,14 +31,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
   
   // pegando parametros da request com (request), recurso do proprio Fastify
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
 
     // Validação para a variavel id que será recebida com parametro
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
-    // Objeto com parametros da rota, sendo passado para o ZOD fazer as validacoes necessarias
+    // Objeto com parametros da rota, sendo passado para o ZOD fazer as validações necessarias
     const { id } = paramsSchema.parse(request.params)
 
     // Encontra uma memoria ou senão, despara um erro
@@ -38,6 +47,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id
       }
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -56,14 +69,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '37e13105-0cb9-456f-81ef-3495f0ca320e'
+        userId: request.user.sub
       }
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -78,7 +91,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -93,12 +116,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
